prune component
