+++ src/test/scala/nl.knaw.dans.easy.bagstore/BagStorePruneSpec.scala
@@ -21,7 +21,7 @@ import org.apache.commons.io.FileUtils
 
 import scala.util.Success
 
-class BagStorePruneSpec extends BagStoreFixture with BagStorePrune with BagStoreAddComponent {
+class BagStorePruneSpec extends BagStoreFixture with BagStorePruneComponent with BagStoreAddComponent {
   import context._
 
   FileUtils.copyDirectory(Paths.get("src/test/resources/bags/basic-sequence-unpruned").toFile, testDir.toFile)
@@ -30,11 +30,12 @@ class BagStorePruneSpec extends BagStoreFixture with BagStorePrune with BagStore
   private val TEST_BAG_C = testDir.resolve("c")
 
   override val add = new BagStoreAdd {}
+  override val prune = new BagStorePrune {}
 
   "prune" should "change files present in ref-bags to fetch.txt entries" in {
     val tryA = add.add(TEST_BAG_A)
     tryA shouldBe a[Success[_]]
-    prune(TEST_BAG_B, tryA.get) shouldBe a[Success[_]]
+    prune.prune(TEST_BAG_B, tryA.get) shouldBe a[Success[_]]
 
     /*
      * Now follow checks on the content of of the ingested bags. Each file should be EITHER actually present in the
@@ -90,7 +91,7 @@ class BagStorePruneSpec extends BagStoreFixture with BagStorePrune with BagStore
     val tryB = add.add(TEST_BAG_B)
     tryB shouldBe a[Success[_]]
 
-    prune(TEST_BAG_C, tryA.get, tryB.get)
+    prune.prune(TEST_BAG_C, tryA.get, tryB.get)
 
     /*
      * Check bag C
